removed the AbstractRepositoryCache and moved all functionality to the AionRepositoryCache
