+++ modAionImpl/src/org/aion/zero/impl/db/AionRepositoryCache.java
@@ -1,18 +1,45 @@
 package org.aion.zero.impl.db;
 
+import static org.aion.crypto.HashUtil.h256;
+import static org.aion.util.bytes.ByteUtil.EMPTY_BYTE_ARRAY;
+
+import java.math.BigInteger;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.locks.ReadWriteLock;
+import java.util.concurrent.locks.ReentrantReadWriteLock;
 import org.aion.interfaces.db.ByteArrayKeyValueStore;
 import org.aion.interfaces.db.ContractDetails;
 import org.aion.interfaces.db.Repository;
 import org.aion.interfaces.db.RepositoryCache;
+import org.aion.log.AionLoggerFactory;
+import org.aion.log.LogEnum;
 import org.aion.mcf.core.AccountState;
 import org.aion.mcf.db.IBlockStoreBase;
 import org.aion.types.Address;
+import org.aion.types.ByteArrayWrapper;
+import org.slf4j.Logger;
 
-public class AionRepositoryCache extends AbstractRepositoryCache<IBlockStoreBase<?, ?>> {
+public class AionRepositoryCache implements RepositoryCache<AccountState, IBlockStoreBase<?, ?>> {
+
+    // Logger
+    protected static final Logger LOG = AionLoggerFactory.getLogger(LogEnum.DB.name());
+
+    /** the repository being tracked */
+    protected Repository<AccountState, IBlockStoreBase<?, ?>> repository;
+
+    /** local accounts cache */
+    protected Map<Address, AccountState> cachedAccounts;
+
+    protected ReadWriteLock lockAccounts = new ReentrantReadWriteLock();
+    /** local contract details cache */
+    protected Map<Address, ContractDetails> cachedDetails;
+
+    protected ReadWriteLock lockDetails = new ReentrantReadWriteLock();
 
     public AionRepositoryCache(final Repository trackedRepository) {
         this.repository = trackedRepository;
@@ -25,22 +52,366 @@ public class AionRepositoryCache extends AbstractRepositoryCache<IBlockStoreBase
         return new AionRepositoryCache(this);
     }
 
+    @Override
+    public AccountState createAccount(Address address) {
+        fullyWriteLock();
+        try {
+            AccountState accountState = new AccountState();
+            cachedAccounts.put(address, accountState);
+
+            // TODO: unify contract details initialization from Impl and Track
+            ContractDetails contractDetails = new ContractDetailsCacheImpl(null);
+            // TODO: refactor to use makeDirty() from AbstractState
+            contractDetails.setDirty(true);
+            cachedDetails.put(address, contractDetails);
+
+            return accountState;
+        } finally {
+            fullyWriteUnlock();
+        }
+    }
+
+    /**
+     * Retrieves the current state of the account associated with the given address.
+     *
+     * @param address the address of the account of interest
+     * @return a {@link AccountState} object representing the account state as is stored in the
+     *     database or cache
+     * @implNote If there is no account associated with the given address, it will create it.
+     */
+    @Override
+    public AccountState getAccountState(Address address) {
+        lockAccounts.readLock().lock();
+
+        try {
+            // check if the account is cached locally
+            AccountState accountState = this.cachedAccounts.get(address);
+
+            // when the account is not cached load it from the repository
+            if (accountState == null) {
+                // must unlock to perform write operation from loadAccountState(address)
+                lockAccounts.readLock().unlock();
+                loadAccountState(address);
+                lockAccounts.readLock().lock();
+                accountState = this.cachedAccounts.get(address);
+            }
+
+            return accountState;
+        } finally {
+            try {
+                lockAccounts.readLock().unlock();
+            } catch (Exception e) {
+                // there was nothing to unlock
+            }
+        }
+    }
+
+    public boolean hasAccountState(Address address) {
+        lockAccounts.readLock().lock();
+        try {
+            AccountState accountState = cachedAccounts.get(address);
+
+            if (accountState != null) {
+                // checks that the account is not cached as deleted
+                // TODO: may also need to check if the state is empty
+                return !accountState.isDeleted();
+            } else {
+                // check repository when not cached
+                return repository.hasAccountState(address);
+            }
+        } finally {
+            lockAccounts.readLock().unlock();
+        }
+    }
+
+    @Override
+    public ContractDetails getContractDetails(Address address) {
+        lockDetails.readLock().lock();
+
+        try {
+            ContractDetails contractDetails = this.cachedDetails.get(address);
+
+            if (contractDetails == null) {
+                // loads the address into cache
+                // must unlock to perform write operation from loadAccountState(address)
+                lockDetails.readLock().unlock();
+                loadAccountState(address);
+                lockDetails.readLock().lock();
+                // retrieves the contract details
+                contractDetails = this.cachedDetails.get(address);
+            }
+
+            return contractDetails;
+        } finally {
+            try {
+                lockDetails.readLock().unlock();
+            } catch (Exception e) {
+                // there was nothing to unlock
+            }
+        }
+    }
+
+    @Override
+    public boolean hasContractDetails(Address address) {
+        lockDetails.readLock().lock();
+
+        try {
+            ContractDetails contractDetails = cachedDetails.get(address);
+
+            if (contractDetails == null) {
+                // ask repository when not cached
+                return repository.hasContractDetails(address);
+            } else {
+                // TODO: may also need to check if the details are empty
+                return !contractDetails.isDeleted();
+            }
+        } finally {
+            lockDetails.readLock().unlock();
+        }
+    }
+
+    /**
+     * @implNote The loaded objects are fresh copies of the locally cached account state and
+     *     contract details.
+     */
+    @Override
+    public void loadAccountState(
+            Address address,
+            Map<Address, AccountState> accounts,
+            Map<Address, ContractDetails> details) {
+        fullyReadLock();
+
+        try {
+            // check if the account is cached locally
+            AccountState accountState = this.cachedAccounts.get(address);
+            ContractDetails contractDetails = this.cachedDetails.get(address);
+
+            // when account not cached load from repository
+            if (accountState == null) {
+                // load directly to the caches given as parameters
+                repository.loadAccountState(address, accounts, details);
+            } else {
+                // copy the objects if they were cached locally
+                accounts.put(address, new AccountState(accountState));
+                details.put(address, new ContractDetailsCacheImpl(contractDetails));
+            }
+        } finally {
+            fullyReadUnlock();
+        }
+    }
+
+    /**
+     * Loads the state of the account into <b>this object' caches</b>. Requires write locks on both
+     * {@link #lockAccounts} and {@link #lockDetails}.
+     *
+     * @implNote If the calling method has acquired a weaker lock, the lock must be released before
+     *     calling this method.
+     * @apiNote If the account was never stored this call will create it.
+     */
+    private void loadAccountState(Address address) {
+        fullyWriteLock();
+        try {
+            repository.loadAccountState(address, this.cachedAccounts, this.cachedDetails);
+        } finally {
+            fullyWriteUnlock();
+        }
+    }
+
+    @Override
+    public void deleteAccount(Address address) {
+        fullyWriteLock();
+        try {
+            getAccountState(address).delete();
+            getContractDetails(address).setDeleted(true);
+        } finally {
+            fullyWriteUnlock();
+        }
+    }
+
+    @Override
+    public BigInteger incrementNonce(Address address) {
+        lockAccounts.writeLock().lock();
+        try {
+            return getAccountState(address).incrementNonce();
+        } finally {
+            lockAccounts.writeLock().unlock();
+        }
+    }
+
+    @Override
+    public BigInteger setNonce(Address address, BigInteger newNonce) {
+        lockAccounts.writeLock().lock();
+        try {
+            return getAccountState(address).setNonce(newNonce);
+        } finally {
+            lockAccounts.writeLock().unlock();
+        }
+    }
+
+    @Override
+    public BigInteger getNonce(Address address) {
+        AccountState accountState = getAccountState(address);
+        // account state can never be null, but may be empty or deleted
+        return (accountState.isEmpty() || accountState.isDeleted())
+                ? BigInteger.ZERO
+                : accountState.getNonce();
+    }
+
+    @Override
+    public BigInteger getBalance(Address address) {
+        AccountState accountState = getAccountState(address);
+        // account state can never be null, but may be empty or deleted
+        return (accountState.isEmpty() || accountState.isDeleted())
+                ? BigInteger.ZERO
+                : accountState.getBalance();
+    }
+
+    @Override
+    public BigInteger addBalance(Address address, BigInteger value) {
+        lockAccounts.writeLock().lock();
+        try {
+            // TODO: where do we ensure that this does not result in a negative value?
+            AccountState accountState = getAccountState(address);
+            return accountState.addToBalance(value);
+        } finally {
+            lockAccounts.writeLock().unlock();
+        }
+    }
+
+    @Override
+    public void saveCode(Address address, byte[] code) {
+        fullyWriteLock();
+        try {
+            // save the code
+            // TODO: why not create contract here directly? also need to check that there is no
+            // preexisting code!
+            ContractDetails contractDetails = getContractDetails(address);
+            contractDetails.setCode(code);
+            // TODO: ensure that setDirty is done by the class itself
+            contractDetails.setDirty(true);
+
+            // update the code hash
+            getAccountState(address).setCodeHash(h256(code));
+        } finally {
+            fullyWriteUnlock();
+        }
+    }
+
+    @Override
+    public byte[] getCode(Address address) {
+        if (!hasAccountState(address)) {
+            return EMPTY_BYTE_ARRAY;
+        }
+
+        byte[] codeHash = getAccountState(address).getCodeHash();
+
+        // TODO: why use codeHash here? may require refactoring
+        return getContractDetails(address).getCode(codeHash);
+    }
+
+    @Override
+    public void addStorageRow(Address address, ByteArrayWrapper key, ByteArrayWrapper value) {
+        lockDetails.writeLock().lock();
+        try {
+            getContractDetails(address).put(key, value);
+        } finally {
+            lockDetails.writeLock().unlock();
+        }
+    }
+
+    @Override
+    public void removeStorageRow(Address address, ByteArrayWrapper key) {
+        lockDetails.writeLock().lock();
+        try {
+            getContractDetails(address).delete(key);
+        } finally {
+            lockDetails.writeLock().unlock();
+        }
+    }
+
+    @Override
+    public ByteArrayWrapper getStorageValue(Address address, ByteArrayWrapper key) {
+        return getContractDetails(address).get(key);
+    }
+
+    @Override
+    public Map<ByteArrayWrapper, ByteArrayWrapper> getStorage(
+            Address address, Collection<ByteArrayWrapper> keys) {
+        ContractDetails details = getContractDetails(address);
+        return (details == null) ? Collections.emptyMap() : details.getStorage(keys);
+    }
+
+    @Override
+    public void rollback() {
+        fullyWriteLock();
+        try {
+            cachedAccounts.clear();
+            cachedDetails.clear();
+        } finally {
+            fullyWriteUnlock();
+        }
+    }
+
+    @Override
+    public Repository getSnapshotTo(byte[] root) {
+        return repository.getSnapshotTo(root);
+    }
+
+    // This method was originally disabled because changes to the blockstore
+    // can not be reverted. The reason for re-enabling it is that we have no way
+    // to
+    // get a reference of the blockstore without using
+    // NProgInvoke/NcpProgInvoke.
+    @Override
+    public IBlockStoreBase<?, ?> getBlockStore() {
+        return repository.getBlockStore();
+    }
+
+    /** Lock to prevent writing on both accounts and details. */
+    protected void fullyWriteLock() {
+        lockAccounts.writeLock().lock();
+        lockDetails.writeLock().lock();
+    }
+
+    /** Unlock to allow writing on both accounts and details. */
+    protected void fullyWriteUnlock() {
+        lockDetails.writeLock().unlock();
+        lockAccounts.writeLock().unlock();
+    }
+
+    /** Lock for reading both accounts and details. */
+    protected void fullyReadLock() {
+        lockAccounts.readLock().lock();
+        lockDetails.readLock().lock();
+    }
+
+    /** Unlock reading for both accounts and details. */
+    protected void fullyReadUnlock() {
+        lockDetails.readLock().unlock();
+        lockAccounts.readLock().unlock();
+    }
+
+    @Override
+    public boolean isSnapshot() {
+        return repository.isSnapshot();
+    }
+
     /**
      * Flushes its state to other in such a manner that other receives sufficiently deep copies of
      * its {@link AccountState} and {@link ContractDetails} objects.
      *
-     * If {@code clearStateAfterFlush == true} then this repository's state will be completely
+     * <p>If {@code clearStateAfterFlush == true} then this repository's state will be completely
      * cleared after this method returns, otherwise it will retain all of its state.
      *
-     * A "sufficiently deep copy" is an imperfect deep copy (some original object references get
+     * <p>A "sufficiently deep copy" is an imperfect deep copy (some original object references get
      * leaked) but such that for all conceivable use cases these imperfections should go unnoticed.
      * This is because doing something like copying the underlying data store makes no sense, both
      * repositories should be accessing it, and there are some other cases where objects are defined
-     * as type {@link Object} and are cast to their expected types and copied, but will not be copied
-     * if they are not in fact their expected types. This is something to be aware of. Most of the
-     * imperfection results from the inability to copy
-     * {@link ByteArrayKeyValueStore} and {@link org.aion.mcf.trie.SecureTrie}
-     * perfectly or at all (in the case of the former), for the above reasons.
+     * as type {@link Object} and are cast to their expected types and copied, but will not be
+     * copied if they are not in fact their expected types. This is something to be aware of. Most
+     * of the imperfection results from the inability to copy {@link ByteArrayKeyValueStore} and
+     * {@link org.aion.mcf.trie.SecureTrie} perfectly or at all (in the case of the former), for the
+     * above reasons.
      *
      * @param other The repository that will consume the state of this repository.
      * @param clearStateAfterFlush True if this repository should clear its state after flushing.
@@ -69,7 +440,7 @@ public class AionRepositoryCache extends AbstractRepositoryCache<IBlockStoreBase
                     contractDetailsCache.commit();
 
                     if (contractDetailsCache.origContract == null
-                        && other.hasContractDetails(entry.getKey())) {
+                            && other.hasContractDetails(entry.getKey())) {
                         // in forked block the contract account might not exist thus
                         // it is created without
                         // origin, but on the main chain details can contain data
@@ -77,7 +448,7 @@ public class AionRepositoryCache extends AbstractRepositoryCache<IBlockStoreBase
                         // into a single storage trie so both branches with
                         // different stateRoots are valid
                         contractDetailsCache.origContract =
-                            other.getContractDetails(entry.getKey()).copy();
+                                other.getContractDetails(entry.getKey()).copy();
                         contractDetailsCache.commit();
                     }
                 }
@@ -118,7 +489,7 @@ public class AionRepositoryCache extends AbstractRepositoryCache<IBlockStoreBase
                     contractDetailsCache.commit();
 
                     if (contractDetailsCache.origContract == null
-                        && other.hasContractDetails(entry.getKey())) {
+                            && other.hasContractDetails(entry.getKey())) {
                         // in forked block the contract account might not exist thus
                         // it is created without
                         // origin, but on the main chain details can contain data
@@ -126,7 +497,7 @@ public class AionRepositoryCache extends AbstractRepositoryCache<IBlockStoreBase
                         // into a single storage trie so both branches with
                         // different stateRoots are valid
                         contractDetailsCache.origContract =
-                            other.getContractDetails(entry.getKey());
+                                other.getContractDetails(entry.getKey());
                         contractDetailsCache.commit();
                     }
                 }
